first paragrpah edit idea (see comment)
Here's another idea. I wonder if we need to "define" "accessibility" as "to people with disabilities" in the opening paragraph? -- to give it context for people who land on this page and don't know. (I was interested in the HTML5 WebX course comments that at least 2 people did not know about accessibility!)