UI design phrase
"user interface design" is a phrase in my brain. It reads awkwardly without it. I stumble on it -- even knowing what it is trying to day. (and we previously agreed on user interface design before visual design)

I want to request this one more time -- then I will let it drop. If it's important to you to leave it as is, OK.