spacing & headings
Providing headings is more a writing task -- so I downplayed it from spacing, and changed "provide" to "style", which is more the designing task. [medium]
I think "proximity" is an understandable term even if you don't have the background in the concept. [low]
"content" includes "elements" so don't need both [low]