Update writing.html.erb.md
Tried to be more accurate on the actual requirements, yet simplify and provide examples. Also avoided saying things like "you may be required" or such, as this does not really help in telling people what they need to do. Not sure we need to define what captions and audio descriptions are in more detail - we link the people to the WCAG requirements anyway, so they can find out more there.