Update index.html.erb.md
I have tweaked the introductory sentence ever so slightly, I have also added a second sentence to try to respond to @nitedog's [strong] comment. There is a little challenge here as @nitedog has a strong comment that @slhenry has an opposing medium-strong comment against it.

On balance I think it is important to highlight that these tips are not comprehensive. I can envisage concerns being raised if this is not made quite clear.