Null alt ref in Write meaningful text alternatives
This page is great!  But <q>no need to write alternative text</q> seems like an invitation to skip the alt attribute.  But how to convey <q>null alt</q> or the requirement that alternative text be <q>implemented in a way that it can be ignored by assistive technology</q> while using plain language?

Current:
> For purely decorative images, there is no need to write alternative text.

Proposed change:
> For purely decorative images that convey no information, do not provide alternative text.